+++ web/src/main/java/com/epam/mediaserver/command/impl/add/GoodAdd.java
@@ -34,7 +34,7 @@ public class GoodAdd implements Command {
 
             ServiceFactory.getOrderUserService().create(user);
 
-            Song song = ServiceFactory.getSongService().getById(Integer.parseInt(id));
+            Song song = ServiceFactory.getSongService().getById(Long.parseLong(id));
 
             ServiceFactory.getOrderUserService().addGoodToOrder(song);
             ServiceFactory.getOrderUserService().getPrice();
